pyproject.toml
[project]
name = "ashworth-engine"
version = "0.1.0"
description = "Multi-agent financial intelligence platform built on LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "langgraph>=0.2.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.0.10",
    "supabase>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
    "PyPDF2>=3.0.0",
    "chardet>=5.2.0",
    "python-magic>=0.4.27",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "langchain-google-genai>=2.1.10",
    "ollama>=0.5.3",
    "click>=8.2.1",
    "tabulate>=0.9.0",
    "aiohttp",
    "aiofiles",
    "httpx>=0.28.1",
    "asyncpg>=0.30.0",
    "openai>=1.102.0",
    "pydantic-settings>=2.10.1",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
